requirements.txt
# Logging 
matplotlib 
tqdm
wandb
tensorboardX

# Config system
hydra-core

# Common libraries
numpy